+++ src/org/jpc/emulator/execution/opcodes/rm/cmp_AL_Ib.java
@@ -17,8 +17,8 @@ public class cmp_AL_Ib extends Executable
 
     public Branch execute(Processor cpu)
     {
-        cpu.flagOp1 = cpu.r_al.get8();
-        cpu.flagOp2 = imm;
+        cpu.flagOp1 = (byte)cpu.r_al.get8();
+        cpu.flagOp2 = (byte)imm;
         cpu.flagResult = (byte)(cpu.flagOp1 - cpu.flagOp2);
         cpu.flagIns = UCodes.SUB8;
         cpu.flagStatus = OSZAPC;
